Refactor `SelectFlowTemplateDialog` for improved modularity (#1196)
